refactor: refactor bucket field add logic, clean types, move components into their main file
